backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
argon2-cffi==25.1.0
PyJWT==2.9.0
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
python-multipart==0.0.22
redis==5.2.1